requirements.txt
beautifulsoup4==4.4.1
docopt==0.6.2
docutils==0.12
google-api-python-client==1.5.0
httplib2==0.9.2
lockfile==0.12.2
logutils==0.3.3
Mako==1.0.4
MarkupSafe==0.23
notify2==0.3
oauth2client==2.0.1
pafy==0.5.0
pecan==1.0.5
pluggy==0.3.1
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyHamcrest==1.9.0
python-daemon==2.1.1
python-vlc==1.1.2
requests==2.9.1
rsa==3.4
simplejson==3.8.2
six==1.10.0
tox==2.3.1
uritemplate==0.6
virtualenv==15.0.0
vk==2.0.2
waitress==0.8.10
WebOb==1.6.0
WebTest==2.0.20
youtube-dl==2016.3.18
